+++ app/src/main/java/com/mobdeve/s15/group8/mobdeve_mp/controller/activities/ProfileActivity.kt
@@ -140,7 +140,7 @@ class ProfileActivity :
         }
     }
 
-    override fun onFeedbackContinue(dialog: DialogFragment, feedbackRating: Float, feedbackComment: String) {
+    override fun onFeedbackContinue(dialog: DialogFragment, feedbackRating: Int, feedbackComment: String) {
         mEditor.putInt(
             getString(R.string.SP_FEED_TIME_KEY),
             TimeUnit.HOURS.convert(Date().time, TimeUnit.MILLISECONDS).toInt()
